A first working implementation of a visual debugger
It is missing a separate window, which shows all local and global variables. Currently, state is printed to the immediate frame since I do not know how to properly setup such a frame. Also, I do not know how to make the code frame non-editable so that code cannot be visually modified while debugging. I guess it could be done by make the frame not focusable for time of debugging but I could not see in the docs how to make that work

Note, The visual debugger is currently only working on platforms on which pexpect is working. Windows support has to be added.